pyproject.toml
[tool.poetry]
name = "my-fastapi-app"
version = "0.1.0"
description = ""
authors = ["tyavkd62 <107627503+tyavkd62@users.noreply.github.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
py-ulid = "^1.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
pymysql = "^1.1.1"
dependency-injector = "^4.46.0"
bcrypt = "4.0.1"
pydantic = {extras = ["email"], version = "^2.11.3"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
celery = "^5.5.2"
requests = "^2.32.3"
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
freezegun = "^1.5.1"

[tool.poetry.group.dev.dependencies]
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"